Factor out process creation code into the platform layer.
At the same time, move platform abstraction code into a separate
subdirectory.

The process-creation code is still very POSIX-specific.  It's not clear
how much code we are going to be able to share between POSIX and
non-POSIX systems and so for now I am focusing on factoring out the
variations between POSIX systems, rather than building a completely
portable abstraction layer.  On Windows, for example, we won't inherit
handles, we'll use DuplicateHandle to add handles to the child process.